Strategy pattern for commands + independent tests
