+++ src/main/java/com/example/happybirthdaybot/common/Answers.java
@@ -32,5 +32,6 @@ public interface Answers {
     String FRIEND_COMMAND = "Вы можете добавить конкретного пользователя в свой список, чтобы получать напоминания независимо от того, состоите ли вы в одном групповом чате!\n\nДля этого просто введите его имя пользователя начиная с символа @";
     String FRIEND_ADDED = "Отлично! Бот обязательно напомнит о Дне Рождения! :)";
     String FRIEND_NOT_FOUND = "К сожалению этот пользователь не зарегистрирован в этом боте :(";
+    String INTERNAL_SERVER_ERROR = "Возникла непредвиденная ошибка сервера :(";
 
 }
