+++ src/main/java/com/example/happybirthdaybot/controllers/ActionController.java
@@ -5,8 +5,8 @@ import com.example.happybirthdaybot.domain.entity.NotificationLevel;
 import com.example.happybirthdaybot.dto.UserDto;
 import com.example.happybirthdaybot.error.ApplicationException;
 import com.example.happybirthdaybot.error.ErrorDescriptions;
-import com.example.happybirthdaybot.service.ChatService;
-import com.example.happybirthdaybot.service.UserService;
+import com.example.happybirthdaybot.service.data.ChatService;
+import com.example.happybirthdaybot.service.data.UserService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
