+++ src/main/java/com/example/happybirthdaybot/error/ErrorDescriptions.java
@@ -19,7 +19,8 @@ public enum ErrorDescriptions {
     APPLICATION_ERROR(Answers.APPLICATION_ERROR, "Invalid input data"),
     JOIN_ERROR(Answers.JOINED_FAILED, "Invalid join data"),
     NO_INFO_ERROR(Answers.NO_INFO, "User not registered"),
-    INVALID_COMMAND_IN_CHAT(Answers.INVALID_COMMAND_IN_CHAT, "Invalid chat command");
+    INVALID_COMMAND_IN_CHAT(Answers.INVALID_COMMAND_IN_CHAT, "Invalid chat command"),
+    HANDLER_NOT_FOUND(Answers.INTERNAL_SERVER_ERROR, "Handler not found");
 
     /**
      * Сообщение пользователю.
