+++ src/test/resources/config/application-test.yaml
@@ -1,0 +1,30 @@
+spring:
+  main:
+    allow-bean-definition-overriding: true
+  application:
+    name: HappyBirthdayBot
+  datasource:
+    driverClassName: org.h2.Driver
+    url: jdbc:h2:./src/test/resources/test_bot;DB_CLOSE_ON_EXIT=TRUE
+    username: sa
+    password:
+  jpa:
+    database-platform: org.hibernate.dialect.H2Dialect
+    hibernate:
+      ddl-auto: update
+  flyway:
+    locations: classpath:db/migration
+  sql:
+    init:
+      platform: h2
+
+
+  rabbitmq:
+    host: ${RABBIT_HOST:localhost}
+    virtualhost: /
+    port: ${RABBIT_PORT:15672}
+    username: guest
+    password: guest
+
+server:
+  port: 8081
